backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
mysql-connector-python==8.0.32
werkzeug==2.2.3
python-dotenv==1.0.0  # 环境变量管理
cryptography==41.0.3   # 用于加密
pytest==7.3.1         # 用于单元测试（可选）
gunicorn==20.1.0      # 用于生产环境部署（可选）
faker==18.10.1        # 用于生成测试数据（可选）
